perception: fix code style except radar/fusion/traffic_light

+++ modules/perception/onboard/subnode_helper.cc
@@ -15,6 +15,7 @@
  *****************************************************************************/
 
 #include <boost/algorithm/string.hpp>
+#include <vector>
 
 #include "modules/common/log.h"
 #include "modules/perception/onboard/subnode_helper.h"
@@ -72,8 +73,8 @@ bool SubnodeHelper::ProduceSharedDataKey(double stamp, const string &device_id,
                                          string *key) {
   char temp[64];
   memset(temp, '\0', sizeof(temp));
-  int ret = snprintf(temp, sizeof(temp), "%ld",
-                     static_cast<long>(stamp * FLAGS_stamp_enlarge_factor));
+  int ret = snprintf(temp, sizeof(temp), "%d",
+                     static_cast<int>(stamp * FLAGS_stamp_enlarge_factor));
   if (ret < 0) {
     AERROR << "Encounter an output error.";
     return false;
